Initial version of the project: launcher script + shell integration files
This commit is the result of squashing the result of the work done by
Niv Sardi <xaiki@evilgiggle.com> and myself in the past weeks, so that
we have a clean and simple initial state.

Note that this is just a baseline, we still need to add icons and i18n
support for the desktop and appdata files, as well as a proper license.

https://phabricator.endlessm.com/T4489